+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ master
 
 1) [BUGFIX] Leave site roots foreach after first hugo enabled site root because content elements are the same for all
     root sites.
+2) [FEATURE] Initial implementation for media export.
 
 0.0.8
 ~~~~~