+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Changelog
 ---------
 
-master
-~~~~~~
+0.0.9
+~~~~~
 
 1) [BUGFIX] Leave site roots foreach after first hugo enabled site root because content elements are the same for all
     root sites.